# central/api/models.py
# generated by datamodel-codegen:
#   filename:  openapi.yaml
#   timestamp: 2026-02-22T21:20:51+00:00

from __future__ import annotations

from enum import Enum
from ipaddress import IPv4Address
from pydantic import BaseModel, Field
from uuid import UUID


class HostStatus(BaseModel):
    ipv4Public: IPv4Address = Field(
        ..., description='External IPv4 address for the host.'
    )


class Format(Enum):
    TEXT = 'text'
    MARKDOWN = 'markdown'


class Notification(BaseModel):
    title: str = Field(..., description="Notification's title")
    content: str = Field(
        ..., description="Notification's body. Emojis are allowed here."
    )
    format: Format | None = Field(
        Format.TEXT,
        description='Notification content format. Defaults to `text` when not specified.\n',
    )


class NotifDeliverySvc(Enum):
    TELEGRAM = 'telegram'
    WEBSOCKET = 'websocket'


class NotifSubscription(BaseModel):
    id: UUID | None = Field(None, description='The UUID4 of the subscription.')
    notifDeliverySvc: NotifDeliverySvc = Field(
        ..., description='Name of service that will receive notifications.'
    )
    password: str = Field(
        ...,
        description='In order to receive notifications, subscription password must\nbe provided.\n',
    )
    servicePayload: str | None = Field(
        None,
        description="This is a flexible field that can be used to transfer additonal\ninfo that might be required by notif delivery service:\n\n1. `telegram`: Init data string provided by telegram to embedded\n   'mini app'.\n\n   Value must be provided as passed by telegram to mini app\n   (without any modification) for validation on backend.\n\n   See: https://core.telegram.org/bots/webapps#validating-data-received-via-the-mini-app\n\n2. `websocket`: No additional info is required\n",
    )